Modificato estratto con descrizione, modificata la home admin e inoltre ogni tag deve avere una descrizione obbligatoria

+++ anteprima-riassunto.php
@@ -23,67 +23,114 @@ include("default-code/caricamento-tags-xml.php");
 	if (isset($_GET['nomeMateria'])) { 
 		//Dopo aver caricato il pdf, compilato correttamente il form e controllato i tag che siano corretti, possiamo aggiornare il DOM
 		if (isset($_SESSION['anteprimaRiassunto'])) {
-			if (!isset($_GET['conferma'])) {
+			if (!isset($_POST['conferma']) && !isset($_GET['conferma'])) {
 			?>
 			<div id="visualizzaRiassunto">
 				<div id="nomeMateria">
 					<b><?= $_SESSION['titoloRiassunto'] ?></b>
 				</div>
-				<div>
-					<?php
-					if (!empty ($_SESSION['descrizioneRiassunto'])) {
-						?>
-						<div style='margin-left: 5px;'><?= nl2br($_SESSION['descrizioneRiassunto']) ?></div> 
-						<br />
+				<form action="anteprima-riassunto.php?nomeMateria=<?php echo $_GET['nomeMateria'] ?>" method="POST">
+					<div>
 						<?php
-					}
-					?>
-					<embed src="<?=$_SESSION['linkDocumentoRiassunto']; ?>" width="100%" height="600px" type='application/pdf'>
-					<br /><br />
-					<table id="tabellaTagEstratti">
-						<tr><th colspan="2">Controllo tag</th></tr>
-						<?php
-						foreach ($_SESSION['tagsRiassuntoNuovo'] as $l => $value) {
-							for ($k=0; $k < $tags->length; $k++) {	
-								//Confrontiamo ogni tag inserito con quelli già presenti in tags.xml
-								if (strcasecmp ($value, $nomeTagText[$k]) == 0 && !empty($estrattoTagText[$k])) {
-									$indiceTrovato[$l] = -1;
-									?>
-									<tr><td><a id='tagAnteprima' href='#'><?= $value ?></a></td><td><?= $estrattoTagText[$k] ?></td></tr>
-									<?php
-									break; //Usiamo il break perchè tanto abbiamo trovato ciò che cercavamo nel nostro for annidato
-								}
-							}
-						}
-						//Confrontiamo ogni tag inserito con quelli già presenti in tags.xml: caso in cui tag non è presente.
-						foreach ($_SESSION['tagsRiassuntoNuovo'] as $p => $value) {
-							for ($k=0; $k < $tags->length; $k++) {	
-								$estrattoTagText[$k] = $estrattoTag[$k]->textContent;
-								//Dobbiamo allora semplicemente creare un tag nuovo
-								if (empty($indiceTrovato[$p]) ||  $indiceTrovato[$p] != -1) {
-									?>
-									<tr><td><a id='tagAnteprima' href='#'><?= $value ?></a></td><td><i>estratto mancante...</i></td></tr>
-									<?php
-									break; //Usiamo il break perchè tanto abbiamo trovato ciò che cercavamo nel nostro for annidato
-								}
-							}
-							if (!$tags->length) { //Caso in cui non esistono ancora tags
-								?>
-								<tr><td><a id='tagAnteprima' href='#'><?= $value ?></a></td><td><i>estratto mancante...</i></td></tr>
-								<?php
-							}
+						if (!empty ($_SESSION['descrizioneRiassunto'])) {
+							?>
+							<div style='margin-left: 5px;'><?= nl2br($_SESSION['descrizioneRiassunto']) ?></div> 
+							<br />
+							<?php
 						}
 						?>
-					</table>
-				</div>
-				<div id="pulsantiAnteprima">
-					<a href="anteprima-riassunto.php?nomeMateria=<?php echo $_GET['nomeMateria']."&conferma=0"; ?> " id="tornaAdAggiungiRiassunto">Indietro</a>
-					<a href="anteprima-riassunto.php?nomeMateria=<?php echo $_GET['nomeMateria']."&conferma=1"; ?>" id="confermaRiassunto">Conferma riassunto</a>	
-				<div>	
+						<embed src="<?=$_SESSION['linkDocumentoRiassunto']; ?>" width="100%" height="600px" type='application/pdf'>
+						<br /><br />
+						<table id="tabellaTagDescrizioni">
+							<tr>
+								<th colspan="2">Controllo tag</th>
+							</tr>
+							<?php
+							foreach ($_SESSION['tagsRiassuntoNuovo'] as $l => $value) {
+								for ($k=0; $k < $tags->length; $k++) {	
+									//Confrontiamo ogni tag inserito con quelli già presenti in tags.xml
+									if (strcasecmp ($value, $nomeTagText[$k]) == 0 && !empty($descrizioneTagText[$k])) {
+										$indiceTrovato[$l] = -1;
+										?>
+										<tr><td><a id='tagAnteprima' href='#'><?= $value ?></a></td><td><?= $descrizioneTagText[$k] ?></td></tr>
+										<?php
+										break; //Usiamo il break perchè tanto abbiamo trovato ciò che cercavamo nel nostro for annidato
+									}
+								}
+							}
+							//Confrontiamo ogni tag inserito con quelli già presenti in tags.xml: caso in cui tag non è presente.
+							foreach ($_SESSION['tagsRiassuntoNuovo'] as $p => $value) {
+								for ($k=0; $k < $tags->length; $k++) {	
+									$descrizioneTagText[$k] = $descrizioneTag[$k]->textContent;
+									//Dobbiamo allora semplicemente creare un tag nuovo
+									if (empty($indiceTrovato[$p]) ||  $indiceTrovato[$p] != -1) {
+										?>
+										<tr>
+											<td>
+												<a id='tagAnteprima' href='#'><?= $value ?></a>
+											</td>
+											<td>
+												<textarea name="aggiungiDescrizione[<?= $p ?>]" placeholder="Aggiungere una descrizione al tag che si sta creando (max 500 caratteri)."><?php if (isset($_SESSION['aggiungiDescrizione'])) { echo $_SESSION['aggiungiDescrizione'][$p]; } ?></textarea>
+											</td>
+										</tr>
+										<?php
+										break; //Usiamo il break perchè tanto abbiamo trovato ciò che cercavamo nel nostro for annidato
+									}
+								}
+								if (!$tags->length) { //Caso in cui non esistono ancora tags
+									?>
+									<tr>
+										<td>
+											<a id='tagAnteprima' href='#'><?= $value ?>
+										</a>
+									</td>
+									<td>
+										<i>descrizione mancante...</i>
+									</td>
+								</tr>
+									<?php
+								}
+							}
+							?>
+						</table>
+					</div>
+					<div id="pulsantiAnteprima">
+						<a href="anteprima-riassunto.php?nomeMateria=<?php echo $_GET['nomeMateria']."&conferma=0"; ?> " id="tornaAdAggiungiRiassunto">Indietro</a>
+						<input type="submit" name="conferma" value="Conferma riassunto" id="confermaRiassunto"/>	
+					<div>
+				</form>	
 			</div>
 			<?php
 			}
-			else if ($_GET['conferma'] == 1) {
+			else if (isset($_POST['conferma'])) {
+				//Carichiamo in un'array di sessione tutti i valori precedentemente immessi, 
+				//casomai per errore tornassimo alla pagina precedente
+				foreach ($_POST['aggiungiDescrizione'] as $key => $value) {
+					$_SESSION['aggiungiDescrizione'][$key] = $value;
+				}
+
+				$errore = false;
+				foreach ($_POST['aggiungiDescrizione'] as $value) {
+					if (strlen($value) > 500 || strlen($value) == 0) {
+						$errore = true;
+						break;
+					}
+				}
+				if ($errore) {
+					?>
+					<div id='message'>
+						<img src="images/iconMessage.png">
+						<div>
+							<strong>La descrizione di un tag non può essere lunga più di 500 caratteri oppure vuota.</strong>
+							<br />
+							Ti stiamo reindirizzando...
+						</div>
+					</div>
+					<?php
+					header("refresh:3; url=anteprima-riassunto.php?nomeMateria=".$_GET['nomeMateria']."");
+					exit();
+				}
+				
 				//DA QUI IN AVANTI SI AGGIORNA IL DOM
 								
 				include("default-code/caricamento-riassunti-xml.php"); //Prima carichiamo tutti i riassunti
@@ -91,7 +138,7 @@ include("default-code/caricamento-tags-xml.php");
 				//$lastIDRiassunto proviene da caricamento-riassunti-xml.php ed è l'ultimo IDRiassunto presente (o che lo è stato)
 				//(visto che abbiamo sempre aggiunto e non è possibile che non siano in ordine di grandezza)...
 				$newID = $lastIDRiassunto +1;
-
+				/*
 				$newRiassunto = $doc3->createElement("riassunto");
 				$newIDRiassunto = $doc3->createElement("ID", $newID );		
 				$newTitoloRiassunto = $doc3->createElement("titolo", $_SESSION['titoloRiassunto']);
@@ -124,10 +171,10 @@ include("default-code/caricamento-tags-xml.php");
 				$root3->appendChild($newRiassunto); //Va fatto con appendChild altrimenti potrebbe creare problemi...
 				$path3 = dirname(__FILE__)."/xml-schema/riassunti.xml";
 				$doc3->save($path3);
-				/***/
+				//////////////
 
 
-				/* AGGIORNIAMO IL FILE STUDENTI.XML (solamente riassunti->creati e coins) */ 
+				//AGGIORNIAMO IL FILE STUDENTI.XML (solamente riassunti->creati e coins)
 				
 				//Aggiungiamo 1 ai coin presenti solo se il riassunto è pubblico
 				if (!strcasecmp ($_SESSION['condivisioneRiassunto'], "pubblico")) {
@@ -146,9 +193,9 @@ include("default-code/caricamento-tags-xml.php");
 				$riassuntiCreati->insertBefore($newRiassuntoIDCreato);
 				$path = dirname(__FILE__)."/xml-schema/studenti.xml"; 
 				$doc->save($path);
-				/***/
-
-				/* AGGIORNIAMO IL FILE TAGS.XML */
+				////////////////
+*/
+				//AGGIORNIAMO IL FILE TAGS.XML
 				for ($k=0; $k < $tags->length; $k++) {
 					$tag = $tags->item($k);						
 					//$l è l'indice del tag, tra quelli inseriti. 
@@ -171,11 +218,11 @@ include("default-code/caricamento-tags-xml.php");
 					if (empty($indiceTrovato[$p]) ||  $indiceTrovato[$p] != -1) {
 						$newTag = $doc2->createElement("tag");
 						$newNome = $doc2->createElement("nome", $value);
-						$newEstratto = $doc2->createElement("estratto", "");
+						$newDescrizione = $doc2->createElement("descrizione", $_SESSION['aggiungiDescrizione'][$p]);
 						$newRiassuntoID = $doc2->createElement("riassuntoID", $newID);
 						
 						$newTag->appendChild($newNome);
-						$newTag->appendChild($newEstratto);
+						$newTag->appendChild($newDescrizione);
 						$newTag->appendChild($newRiassuntoID);	
 
 						$root2->appendChild($newTag);
@@ -184,7 +231,7 @@ include("default-code/caricamento-tags-xml.php");
 
 				$path2 = dirname(__FILE__)."/xml-schema/tags.xml"; //Troviamo un percorso assoluto al file xml di riferimento
 				$doc2->save($path2); //Sovrascriviamolo
-				/***/
+				///////////////////
 
 				
 				unset($_SESSION['nowDate']);
@@ -197,10 +244,14 @@ include("default-code/caricamento-tags-xml.php");
 				unset($_SESSION['aggiungiRiassunto']); //Bisogna fare unset dopo aver aggiornato il DOM
 				unset($_SESSION['aggiungiRiassunto']);
 				unset($_SESSION['anteprimaRiassunto']);
+				/*
 				header('Location: visualizza-riassunto.php?IDRiassunto='.$newID);
-			}
-			else { //Se conferma c'è ma è diversa da 1 (ovvero 0)
+				*/
+				exit();
+			} 
+			else if ($_GET['conferma'] == 0) { //Se conferma c'è ma è diversa da 1 (ovvero 0)
 				unlink ($_SESSION['linkDocumentoRiassunto']);
+				unset($_SESSION['aggiungiDescrizione']);
 				header("Location: aggiungi-riassunto.php?nomeMateria=".$_GET['nomeMateria']."");
 				exit();
 			}
